+++ src/ui/dialogs/integerRangeDialog.h
@@ -1,0 +1,40 @@
+/*
+ * PeTrack - Software for tracking pedestrians movement in videos
+ * Copyright (C) 2026 Forschungszentrum Jülich GmbH, IAS-7
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#ifndef INTEGER_RANGE_DIALOG_H
+#define INTEGER_RANGE_DIALOG_H
+
+#include <QDialog>
+
+class QSpinBox;
+
+class IntegerRangeDialog : public QDialog
+{
+    Q_OBJECT
+public:
+    explicit IntegerRangeDialog(int min, int max, const QString &title, QWidget *parent = nullptr);
+
+    int min() const;
+    int max() const;
+
+private:
+    QSpinBox *mSpin1;
+    QSpinBox *mSpin2;
+};
+
+#endif